# api/openapi.yaml
openapi: 3.1.0
info:
  title: YuanCI API
  version: 0.1.0
  description: >-
    Protected browser API contract. The NewAuthenticated handler is implemented
    and integration-tested; the executable can opt into authenticated preview
    with GitHub login, separate from insecure evaluation. There is no arbitrary
    session issuer. Mutation requests require an exact configured HTTPS Origin and a
    session-bound X-CSRF-Token. Evaluation compatibility is not production auth.
servers:
  - url: /api/v1
paths:
...
  /auth/status:
    get:
      operationId: getAuthenticationStatus
      security: []
      responses:
        '200':
          description: Public state, never credentials or account identifiers
          content:
            application/json:
              schema:
                type: object
                properties:
                  mode: { type: string, enum: [evaluation, file, managed] }
                  initialized: { type: boolean }
                  configured: { type: boolean }
                  callback_url: { type: string }
        '503': { description: Status unavailable; never fall back to evaluation }
...
  /pipelines/validate:
    post:
      operationId: validatePipeline
      parameters:
        - { $ref: '#/components/parameters/BrowserOrigin' }
        - { $ref: '#/components/parameters/CSRFToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [content]
              properties: { content: { type: string, maxLength: 1048576 } }
      responses:
        '200': { description: Valid compiled pipeline }
        '422': { description: Pipeline validation errors }
  /runs:
    get:
      operationId: listRuns
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
        - { name: project_id, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Runs in the authorized project only; unowned evaluation runs are excluded
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Run' }
        '400': { description: Missing or invalid project ID }
        '401': { description: Invalid session }
        '403': { description: Project unavailable or permission denied }
    post:
      operationId: createRun
      parameters:
        - { $ref: '#/components/parameters/BrowserOrigin' }
        - { $ref: '#/components/parameters/CSRFToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [pipeline, project_id]
              properties:
                project_id: { type: string, format: uuid }
                pipeline: { type: string }
                event: { type: string, enum: [manual], default: manual }
                ref: { type: string }
                commit_sha: { type: string }
      responses:
        '201': { description: Run accepted and transactionally queued }
        '401': { description: Invalid session }
        '403': { description: Origin, CSRF or scoped permission denied }
        '409': { description: Deployment commit already registered for this repository and environment }
        '422': { description: Pipeline validation errors or deployment requires a repository and immutable commit SHA }
components:
...
  parameters:
...
    BrowserOrigin:
      name: Origin
      in: header
      required: true
      schema: { type: string }
      description: Exact configured HTTPS public origin; forwarded Host headers are not authority
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
      description: Session-bound value returned by GET /session
  schemas:
...
    Run:
      type: object
      required: [id, pipeline_name, event, status, config_sha256, created_at]
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        created_by: { type: string, format: uuid }
        pipeline_name: { type: string }
        event: { type: string }
        ref: { type: string }
        commit_sha: { type: string }
        status:
          type: string
          enum: [queued, running, waiting_approval, succeeded, failed, canceled]
        config_sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
        plan: { type: object }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
  securitySchemes:
    setupSession: { type: apiKey, in: cookie, name: __Host-yuanci_setup }
    cookieSession: { type: apiKey, in: cookie, name: __Host-yuanci_session }
security:
  - cookieSession: []